+++ R/mitre.R
@@ -20,11 +20,9 @@ getLatestDataSet <- function(verbose = FALSE) {
 
 #' ETL process for all standards, it also create a list of nodes and edges
 #' representing the relationships between standard objects. It needs raw files
-#' downloaded from official MITRE repositories stored in a folder named "data-raw";
-#' set downloadLatest parameter to TRUE and the function will create it for you.
+#' pre-downloaded from official MITRE repositories stored in a folder named "data-raw".
 #'
 #' @param verbose default is TRUE
-#' @param downloadLatest default as FALSE, set to TRUE to download latest raw source files from official MITRE repositories
 #'
 #' @return list of two data frames: nodes and edges
 #' @export
@@ -33,8 +31,8 @@ getLatestDataSet <- function(verbose = FALSE) {
 #' \dontrun{
 #' mitredata <- mitre::parseRawData(TRUE)
 #' }
-parseRawData <- function(verbose = FALSE, downloadLatest = TRUE) {
-  if (downloadLatest) downloadRawData(verbose)
+parseRawData <- function(verbose = FALSE) {
+  getRawData(verbose)
   if (!dir.exists("data-raw")) stop('Please, set downloadLatest to TRUE or ensure that your working directory contains a folder named "data-raw" with MITRE raw files.', call. = FALSE)
 
   if (verbose) print(paste("[#][SHIELD] Start ETL process."))
@@ -101,6 +99,81 @@ parseRawData <- function(verbose = FALSE, downloadLatest = TRUE) {
   return(mitre.data)
 }
 
+getRawData <- function(verbose = FALSE) {
+  if (verbose) print(paste("[#][MITRE] Start Download process."))
+  # Create "data-raw" folder
+  if (!dir.exists("data-raw")) dir.create("data-raw")
+
+  # SHIELD
+  if (verbose) print(paste("[-][SHIELD] Download Tactics ..."))
+  tactics_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-tactics.json"
+  utils::download.file(url = tactics_url, destfile = "data-raw/shield-tactics.json", quiet = T)
+  if (verbose) print(paste("[-][SHIELD] Download Techniques ..."))
+  tech_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-techniques.json"
+  utils::download.file(url = tech_url, destfile = "data-raw/shield-techniques.json", quiet = T)
+  if (verbose) print(paste("[-][SHIELD] Download Opportunities ..."))
+  opport_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-opportunities.json"
+  utils::download.file(url = opport_url, destfile = "data-raw/shield-opportunities.json", quiet = T)
+  if (verbose) print(paste("[-][SHIELD] Download Procedures ..."))
+  proced_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-procedures.json"
+  utils::download.file(url = proced_url, destfile = "data-raw/shield-procedures.json", quiet = T)
+  if (verbose) print(paste("[-][SHIELD] Download Use cases ..."))
+  usecase_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-use_cases.json"
+  utils::download.file(url = usecase_url, destfile = "data-raw/shield-use_cases.json", quiet = T)
+  if (verbose) print(paste("[-][SHIELD] Download Tactic details ..."))
+  tact_det_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-tactic_details.json"
+  utils::download.file(url = tact_det_url, destfile = "data-raw/shield-tactic_details.json", quiet = T)
+  if (verbose) print(paste("[-][SHIELD] Download Technique details ..."))
+  tech_det_url <- "https://raw.githubusercontent.com/motherhack3r/mitre-datasets/master/latest/data-raw/shield-technique_details.json"
+  utils::download.file(url = tech_det_url, destfile = "data-raw/shield-technique_details.json", quiet = T)
+
+  # CVE
+  for (year in 2002:strftime(Sys.Date(), "%Y")) {
+    if (verbose) print(paste0("[-][CVE] Download ", year," CVEs ..."))
+    cve_url <- paste0("https://github.com/motherhack3r/mitre-datasets/raw/master/latest/data-raw/cve-", year,".json.gz")
+    utils::download.file(url = cve_url, destfile = paste0("data-raw/cve-", year,".json.gz"), quiet = T)
+  }
+
+  # CWE
+  if (verbose) print(paste("[-][CWE] Download latest XML definitions ..."))
+  cwe.url  <- "https://github.com/motherhack3r/mitre-datasets/raw/master/latest/data-raw/cwe-mitre.xml.zip"
+  utils::download.file(url = cwe.url, destfile = "data-raw/cwe-mitre.xml.zip", quiet = T)
+  utils::unzip(zipfile = paste0("data-raw/cwe-mitre.xml.zip"),
+               exdir = paste0("data-raw"),
+               overwrite = T)
+  unlink("data-raw/cwe-mitre.xml.zip")
+
+  # CPE
+  if (verbose) print(paste("[-][CPE] Download latest XML definitions ..."))
+  cpe.url  <- "https://github.com/motherhack3r/mitre-datasets/raw/master/latest/data-raw/cpe-nist.xml.zip"
+  utils::download.file(url = cpe.url, destfile = "data-raw/cpe-nist.xml.zip", quiet = T)
+  utils::unzip(zipfile = paste0("data-raw/cpe-nist.xml.zip"),
+               exdir = paste0("data-raw"),
+               overwrite = T)
+  unlink("data-raw/cpe-nist.xml.zip")
+  if (verbose) print(paste("[-][CPE] Download latest JSON definitions ..."))
+  cpe.url  <- "https://github.com/motherhack3r/mitre-datasets/raw/master/latest/data-raw/cpe-nist.json.zip"
+  utils::download.file(url = cpe.url, destfile = "data-raw/cpe-nist.json.zip", quiet = T)
+  utils::unzip(zipfile = paste0("data-raw/cpe-nist.json.zip"),
+               exdir = paste0("data-raw"),
+               overwrite = T)
+  unlink("data-raw/cpe-nist.json.zip")
+
+  # CAPEC
+  if (verbose) print(paste("[-][CAPEC] Download latest XML definitions ..."))
+  capec.url  <- "https://github.com/motherhack3r/mitre-datasets/raw/master/latest/data-raw/capec_latest.xml"
+  utils::download.file(url = capec.url, destfile = "data-raw/capec_latest.xml", quiet = T)
+
+  # CTI
+  if (verbose) print(paste("[-][CTI] Download latest YAML definitions ..."))
+  utils::download.file(url = "https://github.com/mitre/cti/archive/master.zip",
+                       destfile = "data-raw/cti.zip", quiet = T)
+  utils::unzip(zipfile = "data-raw/cti.zip", exdir = "data-raw", overwrite = T)
+  unlink("data-raw/cti.zip")
+
+  if (verbose) print(paste("[#][MITRE] End download process."))
+}
+
 #' Download from official sources raw files saving them in [working_directory]/data-raw/
 #'
 #' @param verbose default is FALSE
@@ -152,12 +225,18 @@ downloadRawData <- function(verbose = FALSE) {
   utils::unzip(zipfile = paste0("data-raw/cwe-mitre.xml.zip"),
                exdir = paste0("data-raw"),
                overwrite = T)
+  unlink("data-raw/cwe-mitre.xml.zip")
 
   # CPE
   if (verbose) print(paste("[-][CPE] Download latest XML definitions ..."))
-  cpe.url  <- "http://static.nvd.nist.gov/feeds/xml/cpe/dictionary/official-cpe-dictionary_v2.3.xml.zip"
-  utils::download.file(url = cpe.url, destfile = "data-raw/cpe-mitre.xml.zip", quiet = T)
-  utils::unzip(zipfile = "data-raw/cpe-mitre.xml.zip", exdir = "data-raw", overwrite = T)
+  cpe.url  <- "https://nvd.nist.gov/feeds/xml/cpe/dictionary/official-cpe-dictionary_v2.3.xml.zip"
+  # utils::download.file(url = cpe.url, destfile = "data-raw/cpe-mitre.xml.zip", quiet = T)
+  r <- httr::GET(cpe.url, httr::write_disk("data-raw/cpe-nist.xml.zip", overwrite = TRUE))
+  if (r$status_code == 200) {
+    utils::unzip(zipfile = "data-raw/cpe-nist.xml.zip", exdir = "data-raw", overwrite = T)
+  } else {
+    downloadCPE_API(verbose)
+  }
 
   # CAPEC
   if (verbose) print(paste("[-][CAPEC] Download latest XML definitions ..."))
@@ -169,6 +248,7 @@ downloadRawData <- function(verbose = FALSE) {
   utils::download.file(url = "https://github.com/mitre/cti/archive/master.zip",
                 destfile = "data-raw/cti.zip", quiet = T)
   utils::unzip(zipfile = "data-raw/cti.zip", exdir = "data-raw", overwrite = T)
+  unlink("data-raw/cti.zip")
 
   if (verbose) print(paste("[#][MITRE] End download process."))
 }