+++ NEWS.md
@@ -1,3 +1,8 @@
+# mitre 0.6.0
+* Removed deprecated nodes in mitre network, kept in standards data frames
+* Raw data is downloaded from github repo to avoid official servers rate limitations
+* CPE source changed to JSON. It include more references to CVE.
+
 # mitre 0.5.2
 * New function as_igraph that transform mitre list of nodes and edges to igraph
 * Rscript to easy automate building latest data sets